+++ kotlinpoet/src/jvmTest/kotlin/com/squareup/kotlinpoet/FunSpecTest.kt
@@ -1238,4 +1238,28 @@ class FunSpecTest {
       """.trimMargin(),
     )
   }
+
+  @Test fun importAliasWithNullableReturn() {
+    val packageName = "org.example"
+    val className = ClassName(packageName, "PersonId")
+    val spec = FileSpec.builder("org.example", "SomeFile")
+      .addAliasedImport(className, "PID")
+      .addFunction(
+        FunSpec.builder("pid")
+          .returns(className.copy(nullable = true))
+          .addCode("return %T()", className)
+          .build()
+      )
+      .build()
+    assertThat(spec.toString()).isEqualTo(
+      """
+      |package org.example
+      |
+      |import org.example.PersonId as PID
+      |
+      |public fun pid(): PID? = PID()
+      |
+      """.trimMargin(),
+    )
+  }
 }