+++ docs/changelog.md
@@ -5,6 +5,8 @@ Change Log
 
 * Fix: Support delegates on anonymous classes. (#2034)
 * Fix: `TypeName.MUTABLE_MAP_ENTRY` now correctly uses the `MutableEntry` nested class name. (#2061)
+* Fix: Aliased import ignored for nullable variants of that type. (#2020)
+* Fix: Aliased import may throw java.lang.IllegalArgumentException if nullable and not nullable variants of it are used. (#2021)
 
 
 ## Version 2.0.0